+++ mex/seq_repo/main.py
@@ -129,7 +129,7 @@ def extracted_activity(
 
 
 @asset(group_name="seq_repo")
-def extracted_access_platform(
+def seq_repo_extracted_access_platform(
     unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
     extracted_primary_source_seq_repo: ExtractedPrimarySource,
 ) -> ExtractedAccessPlatform:
@@ -157,6 +157,7 @@ def extracted_distribution(
     extracted_primary_source_seq_repo: ExtractedPrimarySource,
     extracted_access_platform: ExtractedAccessPlatform,
     extracted_organization_rki: ExtractedOrganization,
+    seq_repo_extracted_access_platform: ExtractedAccessPlatform,
 ) -> dict[str, ExtractedDistribution]:
     """Extract distribution from Seq-Repo."""
     settings = SeqRepoSettings.get()
@@ -167,7 +168,7 @@ def extracted_distribution(
     mex_distributions = transform_seq_repo_distribution_to_extracted_distribution(
         seq_repo_latest_source,
         distribution,
-        extracted_access_platform,
+        seq_repo_extracted_access_platform,
         extracted_organization_rki,
         extracted_primary_source_seq_repo,
     )