+++ tests/grippeweb/test_main.py
@@ -3,7 +3,7 @@ import pytest
 from mex.pipeline import run_job_in_process
 
 
-@pytest.mark.usefixtures("mocked_grippeweb", "mocked_wikidata")
+@pytest.mark.usefixtures("mocked_grippeweb", "mocked_ldap", "mocked_wikidata")
 def test_job() -> None:
     result = run_job_in_process("grippeweb")
     assert result.success