+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - entrypoint `all-extractors`: run all extractors
 
+- transform grippeweb resources
+
 ### Changes
 
 - update mex-common to 0.25.1